+++ packages/lib/components/MazLazyImg.vue
@@ -79,7 +79,7 @@
 
 <style lang="postcss" scoped>
   .m-lazy-img-component {
-    @apply maz-relative maz-inline-flex maz-bg-color-lighter maz-flex-center;
+    @apply maz-relative maz-inline-flex maz-flex-center;
 
     &-loader {
       @apply maz-absolute maz-inset-0 maz-hidden maz-flex-center;