+++ CHANGELOG.md
@@ -1,5 +1,67 @@
 # Change Log
 
+## v3.27.4...v3.28.0
+
+[compare changes](https://github.com/LouisMazel/maz-ui/compare/v3.27.4...v3.28.0)
+
+### 🚀 Enhancements
+
+- **maz-ui:** MazTabsBar - fix anchor ([0dc83f8a](https://github.com/LouisMazel/maz-ui/commit/0dc83f8a))
+
+### 💅 Refactors
+
+- **maz-ui:** MazAvatar - remove background color ([cc4f31fb](https://github.com/LouisMazel/maz-ui/commit/cc4f31fb))
+- **maz-ui:** MazImg - remove background color ([1835c419](https://github.com/LouisMazel/maz-ui/commit/1835c419))
+
+### 🏡 Chore
+
+- Update CHANGELOG.md ([be0bbc9e](https://github.com/LouisMazel/maz-ui/commit/be0bbc9e))
+- **release:** Bump version to v3.28.0 ([1940759e](https://github.com/LouisMazel/maz-ui/commit/1940759e))
+
+### ❤️ Contributors
+
+- LouisMazel <me@loicmazuel.com>
+
+## v3.27.3...v3.27.4
+
+[compare changes](https://github.com/LouisMazel/maz-ui/compare/v3.27.3...v3.27.4)
+
+### 🩹 Fixes
+
+- **maz-ui:** MazTabs - fix anchor link ([b39d6551](https://github.com/LouisMazel/maz-ui/commit/b39d6551))
+- **maz-ui:** MazBtn - apply CSS base font-size ([5e5dbc78](https://github.com/LouisMazel/maz-ui/commit/5e5dbc78))
+
+### 🏡 Chore
+
+- Update CHANGELOG.md ([78828259](https://github.com/LouisMazel/maz-ui/commit/78828259))
+- **release:** Bump version to v3.27.4 ([b2f21c3c](https://github.com/LouisMazel/maz-ui/commit/b2f21c3c))
+
+### ❤️ Contributors
+
+- LouisMazel <me@loicmazuel.com>
+
+## v3.27.2...v3.27.3
+
+[compare changes](https://github.com/LouisMazel/maz-ui/compare/v3.27.2...v3.27.3)
+
+### 💅 Refactors
+
+- **maz-ui:** MazAvatar: show initials if caption has multiple words ([33336263](https://github.com/LouisMazel/maz-ui/commit/33336263))
+- **maz-ui:** Remove font-size CSS rules to all components to inherits font-size (md size only) ([94790813](https://github.com/LouisMazel/maz-ui/commit/94790813))
+
+### 🏡 Chore
+
+- Update CHANGELOG.md ([7a6f0e4d](https://github.com/LouisMazel/maz-ui/commit/7a6f0e4d))
+- **release:** Bump version to v3.27.3 ([53db3452](https://github.com/LouisMazel/maz-ui/commit/53db3452))
+
+### ✅ Tests
+
+- **maz-ui:** MazAvatar - fix ([66a54771](https://github.com/LouisMazel/maz-ui/commit/66a54771))
+
+### ❤️ Contributors
+
+- LouisMazel <me@loicmazuel.com>
+
 ## v3.27.1...v3.27.2
 
 [compare changes](https://github.com/LouisMazel/maz-ui/compare/v3.27.1...v3.27.2)